Fixes #10 - Mapping the same param to more than one fader only updates one fader.
This was a bug introduced in Knobbler4, which was a big refactor of
Knobbler3.

The biggest change was moving the javascript out of the 32
bpatchers that handle each mapping slot into a single unified
Javascript object in the top-level patcher. This solved all kinds of
reliability and performance problems. That unified JS was based on the
original JS, where most of the exercise was to adapt the data structures
and methods to operate on arrays of objects rather than just one object.

When changing a parameter value from the TouchOSC interface, an OSC
message is sent from the tablet to the computer, which is received by
the Knobbler4 device. The device processes the message and, in this
case, sets the value of a parameter in the Live Set. That parameter will
emit its value to anything that is listening to value changes, which
includes the mapped parameter slot in Knobbler attached to that
parameter.

To avoid an infinite loop condition, Knobbler implements a timed lock on
paying attention to param value updates on a knob slot that has just
received a value from the tablet. tl;dr - If you move a slider on the
tablet, that slider will not receive param value updates for 300
milliseconds.

In Knobbler3, this lock was global in the code, which was running in 32
instances of the knob slot [bpatcher], so it was local to the slot.

In Knobbler4, there is only one JS instance, and I failed to properly
implement the lock on a per-slot basis. This was the case with four more
locks, though those would not really cause a problem bad enough to open
an issue right away. I fixed all cases, and DRYed up task management /
debouncing logic into a helper function.